+++ tms-service/src/main/java/org/intalio/tempo/workflow/tms/server/TMSServer.java
@@ -77,7 +77,7 @@ public class TMSServer implements ITMSServer {
         this._permissions = permissions;
     }
 
-    public long getHttpTimeout() {
+    public int getHttpTimeout() {
         return _httpTimeout;
     }
